% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.R
\name{division}
\alias{division}
\title{A Division Function}
\usage{
division(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\description{
This function divides one number by another
}
