% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtraction}
\alias{subtraction}
\title{A Subtraction Function}
\usage{
subtraction(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\description{
This function subtracts one number from another
}
